+++ modrinth/CHANGELOG.md
@@ -1,1 +1,1 @@
- - Change namespace initialization (`init`) to `ns`
+ - It is now no longer required to pass a `BlockWithEntity` when registering a `BlockEntity`. Pass a `Block` now.